Fix crash

+++ pynyaata/__init__.py
@@ -108,13 +108,13 @@ def list_animes(url_filters='nyaa,yggtorrent'):
 
 def remove_garbage(link):
     title = link.title
-    folder = title.folder
-    if not len(title.links):
-        db.session.delete(title)
-        db.session.commit()
-    if not len(folder.titles):
-        db.session.delete(folder)
-        db.session.commit()
+    if title:
+        if not len(title.links):
+            db.session.delete(title)
+            db.session.commit()
+        if not len(title.folder.titles):
+            db.session.delete(title.folder)
+            db.session.commit()
 
 
 @app.route('/admin', methods=['GET', 'POST'])
